refactor(ui): extract monolith UI components into modules
This breaks down the massive 1600 line home_screen.dart file. It was getting way too big for new contributors to read through.

I didn't change any logic or rename variables. I just extracted the colors to a theme folder, the math and painters to an animations folder, and the buttons to a widgets folder. The home screen file is now around 400 lines and just handles the layout.
